requirements.txt
pymilvus
timm
einops
ftfy
peft
tqdm
fasthtml
shad4fast
transformers
torchvision
